+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Exercise tracking.
-
+- BMI panel with current data and last update date.
 
 ## 2026-07-17